Add files via upload
The Space Where a Heart Should be, tells Zoran's perspective, where Raevyn was talking to Luci after the change to that one's emotional preservation after the stardust.
Oh How Cruel is The Fate that just Laughs and Watches, is a short one about Raevyn at the end of their lab time, overblotting. And killing all the scientists.